Try to make fast ACUTE code work
