--- tests/crowdsourcing/tasks/fast_acute/test_fast_acute/base__acute_eval_test.grid.winners_as_rows.csv
@@ -1,3 +1,0 @@
-,winner,blender_90m_copy1,blender_90m_copy2
-0,blender_90m_copy1,,0.5
-1,blender_90m_copy2,0.5,
