--- tests/crowdsourcing/tasks/fast_acute/test_fast_acute/base__acute_eval_test.ratings_per_worker.csv
@@ -1,2 +1,0 @@
-,worker,run_id
-0,1,4
