--- tests/crowdsourcing/tasks/fast_acute/test_fast_acute/base__acute_eval_test.significance.csv
@@ -1,2 +1,0 @@
-,question,matchup,model1,numwins1,winrate1,model2,numwins2,winrate2,numratings,sigf,stars,p,agree
-0,Who would you prefer to talk to for a long conversation?,blender_90m_copy1__vs__blender_90m_copy2,blender_90m_copy1,2,0.5,blender_90m_copy2,2,0.5,4,p>.05,,1,
