--- tests/crowdsourcing/tasks/fast_acute/test_fast_acute/q_function__acute_eval_test.grid.csv
@@ -1,3 +1,0 @@
-,loser,blender_90m_copy2,blender_90m_copy1
-0,blender_90m_copy2,,0.75
-1,blender_90m_copy1,0.25,
